test_mgc_pinctrl.sh: New script to test manual gain pin control
This script may be used to test MGC pin control.
By pulsing the control input pins to move the gain indices.
CTRL_IN0 causes the gain index to increase for Rx1 and CTRL_IN1
causes the gain index to decrease for Rx1. Similarly, CTRL_IN2
causes the gain to increase for Rx2 and CTRL_IN3 causes the gain
to decrease for Rx2.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>