adi_update_tools: Shift iio-oscilloscope to cmake build
This patch will use the cmake build option for iio-oscilloscope. It will
fix the issue with deprecated gksudo desktop entry. This has been replaced
with pkexec but only for the cmake build option not for the old make.

Also moving forward cmake will be the default for this tool so it makes
more sense to shift now.

Signed-off-by: Mircea Caprioru <mircea.caprioru@analog.com>