iiod_context: fix hdl sysid with colors
When there are colors codes in the axi sysid strings, the resulting
/etc/libiio.ini is wrongly written. Well, this only happens in the
second time the file is written. If there's no file, it works for the
fist time.

Anyways, this all means that we are unable to get the libiio context
from a remote connection.

In commit 06afede486c3 ("iiod_context.sh: remove non-printable chars in dmesg"),
we tried to remove the specific color codes from the string but that change
is missing to properly escape '\' so that sed actually matches the pattern to
remove. Let's fix that...

Signed-off-by: Nuno Sa <nuno.sa@analog.com>