test_adrv9002.sh: fix bash warning about null bytes
Starting with version 4.4, bash warns about discarded null bytes from
the command substitution output. That was happening when reading
'/sys/firmware/devicetree/base/model' as we have a null byte terminating
the string. Since we are just fine without that byte let's explicitly
remove it with 'tr' so bash does not warn about it.

While at it, replaced 'cat' with '<' as it's faster (not that it matters
much but...)

Signed-off-by: Nuno Sa <nuno.sa@analog.com>