adi_update_tool.sh: set ADI-script default to main
Right now, there are both branches in linux_image-ADI-scripts: 'master'and
'main'. The old version of the script, so what's on old SD card images will
still look for master branch of this repo. The new ones will look for main.
The plan is to keep them in sync for a while (at least until next release)
and afterwards we can delete master. It should not matter which one is the
default branch, since they are in sync.

Signed-off-by: Stefan Raus <Stefan.Raus@analog.com>