adi_update_tools.sh: remove very old releases
We are offering support only for last 2 releases (at the moment this
script was updated - 2021_r1 and 2021_r2). So this commit removed
support for releases older than that. Even if we'll keep them, most
probably they won't work anymore since a lot of things changes on
rootfs side and packages' versions.

Signed-off-by: Stefan Raus <Stefan.Raus@analog.com>