adi_update_boot.sh: keep overlays/overlay_map.dtb
There is a file called 'overlay_map.dtb' in /boot/overlays. If it's
deleted, whenever we use line 'dtoverlay=vc4-kms-v3d' in config.txt,
the bootloader will load 'vc4-kms-v3d' as it is. But if overlay_map.dtb
exists, and booting on RPI4, it will actually use 'vc4-kms-v3d-pi4'.

Actually, there are 2 solutions to fix this problem (no video output
after adi_update_boot.sh is run):

1. Keep overlay_map.dtb and 'dtoverlay=vc4-kms-v3d' in config.txt

2. No overlay_map.dtb and, after loading 'dtoverlay=vc4-kms-v3d' (that
works for RPI 2 and 3, load the pi4 specific one if booting on RPI4:
  [pi4]
  dtoverlay=vc4-kms-v3d-pi4

Signed-off-by: Stefan Raus <Stefan.Raus@analog.com>