test_adrv9002_fh: ensm_mode through device IIO interface
The adrv9002 driver can now be in control of the port enable pins which
means that using the GPIO export interface will fail (pins busy
already). Therefore, if we fail to export the pin, we'll just assume the
driver is in control of it and we'll change ensm_mode as if port enable mode
is set to spi.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>