usb-gadget-service/scripts: handle SDP/ARDZ EEPROMs
Handle reading the EEPROM on SDP/ARDZ boards. These use a different
data format from the FMC boards and need a different parser. There
doesn't seem to be a general purpose tool for this, so add a new
Python script here for that and use that in the iiod_context.sh script
to read the EEPROM and update the /etc/libiio.ini file.

This way, libiio will be able to pick up the hw_mezzanine and hw_serial

Signed-off-by: David Lechner <dlechner@baylibre.com>